# docs/swagger.yaml
basePath: /
definitions:
...
  models.Age:
    properties:
      age:
        type: string
      id:
        type: integer
      posterUrl:
        type: string
    type: object
  models.AllSeries:
    properties:
      description:
        type: string
      director:
        type: string
      id:
        description: Указатель на int
        type: integer
      rating:
        type: integer
      release_year:
        type: integer
      series:
        type: integer
      title:
        type: string
      trailer_url:
        type: string
    type: object
  models.ApiError:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: "OZINSHE\tAPI"
  version: "1.0"
paths:
  /ages:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: List of age
          schema:
            items:
              $ref: '#/definitions/models.Age'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all ages
      tags:
      - ages
    post:
      consumes:
      - application/json
      parameters:
      - description: Age model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Age'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid request age
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: HandleAddAge age
      tags:
      - ages
  /ages/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Ages id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/models.Age'
        "400":
          description: Invalid age Id
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete age
      tags:
      - ages
    get:
      consumes:
      - application/json
      parameters:
      - description: Ages id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/models.Age'
        "400":
          description: Invalid age Id
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Find by id
      tags:
      - ages
    put:
      consumes:
      - application/json
      parameters:
      - description: Age model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Age'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid Age Id
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update age
      tags:
      - ages
  /allseries:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: List of allseries
          schema:
            items:
              $ref: '#/definitions/models.AllSeries'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all allseries
      tags:
      - allseries - это эндпоинты для каждой серии
    post:
      consumes:
      - application/json
      parameters:
      - description: AllSeries model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AllSeries'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid request AllSeries
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create allseries
      tags:
      - allseries - это эндпоинты для каждой серии
  /allseries/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Allseries id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/models.AllSeries'
        "400":
          description: Invalid AllSeries Id
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete allseries
      tags:
      - allseries - это эндпоинты для каждой серии
    get:
      consumes:
      - application/json
      parameters:
      - description: AllSeries id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/models.AllSeries'
        "400":
          description: Invalid allseries id
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Find by id allseries
      tags:
      - allseries - это эндпоинты для каждой серии
    put:
      consumes:
      - application/json
      parameters:
      - description: AllSeries model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AllSeries'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid AllSeries Id
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update allseries
      tags:
      - allseries - это эндпоинты для каждой серии
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"